requirements.txt
torch>=2.0.0
transformers>=4.35.0
langgraph>=0.0.20
langchain>=0.1.0
langchain-community>=0.0.10
duckduckgo-search>=4.1.0
pydantic>=2.5.0
ollama>=0.1.0
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0
typing-extensions>=4.8.0
requests>=2.31.0
openai>=1.0.0
anthropic>=0.18.0
sentence-transformers>=2.2.0
tqdm>=4.66.0

